Replace Plate by Lexical

+++ src/features/post/PostDetail.tsx
@@ -1,4 +1,3 @@
-import { useDebounce, useIsFirstRender } from '@/hooks';
 import TagService from '@/services/tag';
 import { Tag } from '@/types/tag';
 import {
@@ -7,13 +6,13 @@ import {
   WarningOutlined,
 } from '@ant-design/icons';
 import { Button, notification, Select, Typography } from 'antd';
-import CustomPlate from 'components/CustomPlate';
-import { DEFAULT_PLATE_VALUE } from 'components/CustomPlate/utils';
-import { memo, useEffect, useRef, useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
+import { memo, useEffect, useState } from 'react';
 import './Post.less';
 import PostService from './service';
 import { Post } from './types';
+import CustomLexical from 'components/customLexical/CustomLexical';
+import { DEFAULT_VALUE } from 'components/customLexical/utils';
+import { useDebounce } from '@/hooks';
 
 const { Paragraph } = Typography;
 
@@ -28,32 +27,25 @@ const MemoizedPostDetail = memo(function PostDetail({
   onUpdateSuccess,
   onDeleteSuccess,
 }: IProps) {
-  const isFirstRender = useIsFirstRender();
-  const [plateId, setPlateId] = useState(uuidv4());
-  const [content, setContent] = useState(DEFAULT_PLATE_VALUE);
-  const [postTitle, setPostTitle] = useState('');
   const [loading, setLoading] = useState(false);
   const [confirmDelete, setConfirmDelete] = useState(false);
-  const [isUpdated, setIsUpdated] = useState<boolean>(true);
-  const debouncedPost = useDebounce<string>(JSON.stringify(content), 500);
-  const preventUpdate = useRef(false);
   const [listTags, setListTags] = useState<Tag[]>([]);
-  const [selectedTagId, setSelectedTagId] = useState<number>();
+  const [post, setPost] = useState<Post | undefined>();
+  const [lexicalData, setLexicalData] = useState<string | undefined>(
+    JSON.stringify(DEFAULT_VALUE)
+  );
+
+  const debouncePostContent = useDebounce(post?.content, 300);
 
   useEffect(() => {
     (async () => {
       try {
         setLoading(true);
         const res: any = await PostService.detailPost(postId);
-
-        preventUpdate.current = true;
-
         setLoading(false);
         if (res.data && res.data.length === 1) {
-          setContent(JSON.parse(res.data[0].content));
-          setPostTitle(res.data[0].title);
-          setSelectedTagId(res.data[0].tag);
-          setPlateId(uuidv4());
+          setPost(res.data[0]);
+          setLexicalData(res.data[0].content);
         }
       } catch (e) {
         setLoading(false);
@@ -63,21 +55,14 @@ const MemoizedPostDetail = memo(function PostDetail({
 
   const handleUpdate = async () => {
     try {
-      const data: Partial<Post> = {
-        title: postTitle,
-        content: JSON.stringify(content),
-        tag: selectedTagId,
-      };
+      if (!post) return;
       setLoading(true);
-      const res = await PostService.updatePost(postId, data);
-
+      const res = await PostService.updatePost(postId, post);
       setLoading(false);
-      setIsUpdated(true);
       if (res.data && res.data.length === 1) {
         onUpdateSuccess && onUpdateSuccess(res.data[0]);
       }
     } catch (e) {
-      setIsUpdated(false);
       setLoading(false);
       notification.error({ message: 'Error Update Post' });
     }
@@ -85,27 +70,37 @@ const MemoizedPostDetail = memo(function PostDetail({
 
   const handleDelete = async () => {
     try {
+      if (!post) return;
       await PostService.deletePost(postId);
       onDeleteSuccess && onDeleteSuccess(postId);
       notification.success({
-        message: `Delete ${postTitle} successfully`,
+        message: `Delete ${post.title} successfully`,
       });
     } catch (e) {
       notification.error({ message: 'Error Delete Post' });
     }
   };
 
-  const handleChange = (e: any) => {
-    preventUpdate.current = false;
-    setContent(e);
-    setIsUpdated(false);
+  const handleChangeLexical = (value?: string) => {
+    setLexicalData(undefined);
+    setPost({
+      ...post,
+      content: value,
+    } as Post);
   };
+
+  useEffect(() => {
+    if (!debouncePostContent) return;
+
+    handleUpdate();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [debouncePostContent]);
 
   const getListTags = async () => {
     const res = await TagService.listTag();
 
     if (res && res.data) {
-      const newTags: any = res.data.map((i: Tag) => {
+      const newTags: any = (res.data as Tag[]).map((i: Tag) => {
         return {
           label: i.title,
           value: i.id,
@@ -117,26 +112,19 @@ const MemoizedPostDetail = memo(function PostDetail({
   };
 
   const handleChangeTag = (value: any, data: any) => {
-    setSelectedTagId(value);
+    setPost({
+      ...post,
+      tag: data.data.id,
+    } as Post);
   };
 
   useEffect(() => {
     getListTags();
   }, []);
 
-  useEffect(() => {
-    setIsUpdated(true);
-  }, [postId]);
-
-  useEffect(() => {
-    if (preventUpdate.current || isFirstRender) return;
-    handleUpdate();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [debouncedPost]);
-
   return (
     <div className="PostDetail width-100">
-      {isUpdated ? (
+      {!loading ? (
         <Button
           size="small"
           type="primary"
@@ -222,27 +210,23 @@ const MemoizedPostDetail = memo(function PostDetail({
             // icon: <HighlightOutlined />,
             tooltip: 'click to edit text',
             onChange: (text: any) => {
-              setPostTitle(text);
+              setPost({ ...post, title: text } as Post);
             },
             triggerType: ['text'],
           }}
         >
-          {postTitle}
+          {post?.title}
         </Paragraph>
         <Select
           style={{ width: '100px' }}
-          value={selectedTagId}
+          value={post?.tag}
           placeholder="Tags Mode"
           onChange={handleChangeTag}
           options={listTags}
         />
       </div>
       <div style={{ flex: 1, overflow: 'auto' }}>
-        <CustomPlate
-          id={String(plateId)}
-          value={content}
-          onChange={handleChange}
-        />
+        <CustomLexical data={lexicalData} onChange={handleChangeLexical} />
       </div>
     </div>
   );
